added twitter and github to the navbar
